fix: failed replacement machine test
WithError() doesn't automatically set the MachineInfo status to UnReady.
Which is the condition we check on to determine if the replacement is
misbehaving and to later determine if it has errored.

Thus we also need to set WithReady(false).

This closely mimic the real behaviour of the MachineInfo's content
observed in a real cpms cluster.