fix: cluster state validation's: checkNoErrorForReplacements
prior to this fix the check was relying on finding replacement machines
via the updatedMachines() method.
This method though list MachineInfos with machines that have an Spec up-to-date and are Ready.
But in this case we want to find machine that have an up-to-date Spec
and are not-Ready. We need to use pendingMachines() instead.