fix: MachineInfo generation, don't error on unknown index
we don't want to error when an index is unknown, but rather we should
not inject the failure domain into the provider config and carry.
This will allow us to catch the unknown index later on in the reconcile
(at cluster state validation stage), where we'll instead degrade the
operator.