Ensure all machine indexes are represented in the mapping including deleting machines
This changes the way the mapping logic handles indexes that only contain deleting machines.
Rather than ignoring them, we de-prioritise them when matching candidates so that they still
get assigned a failure domain in the output mapping.

If we don't do this and the indexing doesn't match 0, 1, 2, then the index containing
only deleting machines gets assigned a random integer from {0,1,2} rather than the
index currently on the machine.