fix stale cache issue on createMachine
Fixes issue OCPBUGS-4297, where a stale cache of MachineInfos
can lead to the controller not detecting an already existing replacement being present
and creating a subsequent replacement machine.

It fixes the issue by introducing an uncached API client which is later
used to create an uncached machine provider that checks, right before
replacement machine creation (for performance reasons, as we don't want
to always use the expensive uncached client), if a replacement already
exists, avoiding the double replacement problem.