e2e: periodics: remove test with cluster-wide-proxy
following discussions on the 9th Feb 2023 during the Cluster Lifecycle
arch call https://docs.google.com/document/d/1RAkb9QL8Gg7USFiuZi2RpgTZwi-ieTDCdG4YOOkxWaE
we decided that this test is too disruptive for us as the
cluster operators recovery from the cluster-wide-proxy setup takes too
long and break the overall allotted time for this test.
Furthermore we are already testing machine creation under
cluster-wide-proxy within the
machine-api E2Es.

As such we are removing this test and will collaborate with the wider
OpenShift team to resurface a cluster-wide-proxy test for the entire
OpenShift when/if that comes to fruition.