fix: actually return on already-there replacement machine detection
previously, in OCPBUGS-4297 (https://github.com/openshift/cluster-control-plane-machine-set-operator/pull/149) we included
a check on whether a replacement machine already existed and was not previously detected due to potential stale cache.
The check works, and correctly reports a log when a replacement machine is already
present, but we actually forgot to include a return, to stop it from
creating the acutal machine.