Fix MachineNamePrefix periodic on OpenStack
This test assumed that multiple calls to
IncreaseControlPlaneMachineInstanceSize would trigger a rollout every
time. However, this was not true on OpenStack, as flavor names are not
predictable. The OpenStack function assumed it would only ever be called
once, and set a single alternate flavor name passed as an environment
variable. This meant that when called a second time, it did not trigger
a rollout.

In this change we rename the 3 'Increase...InstanceSize' functions to
better reflect their intended function, which is to trigger a rollout.
With this context, we update the OpenStack implementations to add a
random tag instead. This was already the fallback behaviour if no
alternate flavor was passed, and has the additional advantage that it
can be called as many times as required.